create a small powercheck debounced input component
